zakupki-rmksib/requirements.txt
# ============================================
# Bot Framework & Core Dependencies
# ============================================
aiogram>=3.11.0
magic_filter>=0.1.0
python-telegram-bot>=20.0

# ============================================
# Database & ORM
# ============================================
SQLAlchemy>=2.0.30,<3.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
# Раскомментируйте если используете PostgreSQL async:
# asyncpg>=0.29.0

# ============================================
# Web & HTTP Clients
# ============================================
httpx>=0.24.0,<1.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.28.0

# ============================================
# Async & Scheduling
# ============================================
redis>=5.0,<6.0
tenacity>=9.0.0
apscheduler>=3.10.0

# ============================================
# Configuration & Settings
# ============================================
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# ============================================
# Email & Communication
# ============================================
aiosmtplib>=3.0.0
aioimaplib>=1.0.0

# ============================================
# Utilities
# ============================================
orjson>=3.9.0
cryptography>=41.0.0
loguru>=0.7.0

# ============================================
# Document Processing
# ============================================
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# ============================================
# Development Tools
# ============================================
pipdeptree>=2.29.0
